rb5: aosp: fix LUN5 partition scheme
On some (newer?) RB5 devices the LUN5 disk (/dev/sdf) is 32MB, while
it was 128MB on other (older?) devices. When we did the original
partitioning we moved a bunch of partitions which Qualcomm had on LUN0
to LUN5, e.g. ssd, persist, misc, keystore and frp. They were moved
because they seemed unused. However it increased the minimal size for
LUN5 to be around 40MB, which breaks on the newer devices with 32MB
only for LUN5.

But:
* persist was added into LUN0 in
9d5a23e60db7189149ed3f04f94f85a76ab3e21b, but not removed from LUN5.
* misc is a required partition in AOSP vanilla build, so it would be
better to have it on LUN0 than LUN5.
* keystore might be needed, so bring it to LUN0.
* ssd and frp are most likely unused for AOSP builds, but
they are small, so we can keep them on LUN5 without much concern.

Reported-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>