db845c: aosp: partition.xml: Update boot and vendor_boot partition size
Stole some MBs from system, modem and mdtp partitions to bump
up the size of vendor_boot and boot partitions. We need this
to support bigger ramdisks, because we are going to copy a few
firmware files and drop the firmware loading dependency on
userspace helper (FW_LOADER_USER_HELPER_FALLBACK).

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>